[1/10] Replace 'double' with 'float' in OSD code
In the OSD-widget related code replace all (64-bit) double types with (32-bit)
float types. 'float' should be more than accurate enough for what we need. (In
fact openGL uses 32-bit for all coordinate calculations, and there 32-bit is
never a problem, so it certainly won't be a problem for simple 2D
calculations).

On x86 hardware there is not that much advantage in using float instead of
double: a bit less memory usage, maybe slightly more opportunities for SIMD
vectorization, but the actual floating point calculations are about as fast for
double as for float. However on CPUs without hardware floating point support
this can make a difference.

The first patches in this series (replace double->float) also make
preparartions for the later patches (replace scalar->vector).

+++ src/console/OSDImageBasedWidget.cc
@@ -155,17 +155,17 @@ bool OSDImageBasedWidget::hasConstantAlpha() const
 	return constantAlpha(rgba);
 }
 
-double OSDImageBasedWidget::getRecursiveFadeValue() const
+float OSDImageBasedWidget::getRecursiveFadeValue() const
 {
 	return getParent()->getRecursiveFadeValue() * getCurrentFadeValue();
 }
 
 bool OSDImageBasedWidget::isFading() const
 {
-	return (startFadeValue != fadeTarget) && (fadePeriod != 0.0);
+	return (startFadeValue != fadeTarget) && (fadePeriod != 0.0f);
 }
 
-double OSDImageBasedWidget::getCurrentFadeValue() const
+float OSDImageBasedWidget::getCurrentFadeValue() const
 {
 	if (!isFading()) {
 		return startFadeValue;
@@ -173,22 +173,22 @@ double OSDImageBasedWidget::getCurrentFadeValue() const
 	return getCurrentFadeValue(Timer::getTime());
 }
 
-double OSDImageBasedWidget::getCurrentFadeValue(uint64_t now) const
+float OSDImageBasedWidget::getCurrentFadeValue(uint64_t now) const
 {
 	assert(now >= startFadeTime);
 
 	int diff = int(now - startFadeTime); // int should be big enough
-	assert(fadePeriod != 0.0);
-	double delta = diff / (1000000.0 * fadePeriod);
+	assert(fadePeriod != 0.0f);
+	float delta = diff / (1000000.0f * fadePeriod);
 	if (startFadeValue < fadeTarget) {
-		double tmp = startFadeValue + delta;
+		float tmp = startFadeValue + delta;
 		if (tmp >= fadeTarget) {
 			startFadeValue = fadeTarget;
 			return startFadeValue;
 		}
 		return tmp;
 	} else {
-		double tmp = startFadeValue - delta;
+		float tmp = startFadeValue - delta;
 		if (tmp <= fadeTarget) {
 			startFadeValue = fadeTarget;
 			return startFadeValue;
@@ -213,13 +213,13 @@ void OSDImageBasedWidget::invalidateLocal()
 }
 
 void OSDImageBasedWidget::getTransformedXY(const OutputRectangle& output,
-                                           double& outx, double& outy) const
+                                           float& outx, float& outy) const
 {
 	const OSDWidget* parent = getParent();
 	assert(parent);
 	int factor = getScaleFactor(output);
-	double x = factor * getX();
-	double y = factor * getY();
+	float x = factor * getX();
+	float y = factor * getY();
 	parent->transformXY(output, x, y, getRelX(), getRelY(), outx, outy);
 }
 
@@ -274,9 +274,9 @@ void OSDImageBasedWidget::paint(OutputSurface& output, bool openGL)
 
 	byte fadedAlpha = getFadedAlpha();
 	if ((fadedAlpha != 0) && image) {
-		double x, y;
+		float x, y;
 		getTransformedXY(output, x, y);
-		image->draw(output, int(x + 0.5), int(y + 0.5), fadedAlpha);
+		image->draw(output, int(x + 0.5f), int(y + 0.5f), fadedAlpha);
 	}
 	if (isFading()) {
 		gui.refresh();
